[level 0] Title: 두 수의 나눗셈, Time: 0.03 ms, Memory: 71.9 MB -BaekjoonHub

+++ 프로그래머스/0/120806. 두 수의 나눗셈/두 수의 나눗셈.java
@@ -1,5 +1,5 @@
 class Solution {
     public int solution(int num1, int num2) {
-        return (int) ((double) num1 / num2 * 1_000);
+        return (int)(((double) num1 / (double) num2) * 1_000);
     }
 }
